refactor(hyprland): add ImageMagic check for colorpicker

+++ dot_config/hypr/scripts/executable_colorpicker.sh
@@ -5,17 +5,21 @@ if ! command -v hyprpicker >/dev/null 2>&1; then
     exit 1
 fi
 
-color=$(hyprpicker -a)
+if ! command -v magick >/dev/null 2>&1; then
+    hyprpicker -a -n
+else
+    color=$(hyprpicker -a)
 
-if [ "$color" ]; then
-    image=$(mktemp -t colorpicker-XXXXXX.png)
+    if [ "$color" ]; then
+        image=$(mktemp -t colorpicker-XXXXXX.png)
 
-    # Generate preview
-    magick -size 48x48 xc:"$color" "$image"
+        # Generate preview
+        magick -size 48x48 xc:"$color" "$image"
 
-    # Send notification with color
-    notify-send -u low -h string:x-dunst-stack-tag:colorpicker -i "$image" "$color, copied to clipboard."
+        # Send notification with color
+        notify-send -u low -h string:x-dunst-stack-tag:colorpicker -i "$image" "$color, copied to clipboard."
 
-    # Remove temporary preview
-    (sleep 1 && rm -f "$image") &
+        # Remove temporary preview
+        (sleep 1 && rm -f "$image") &
+    fi
 fi
